requirements.txt
pandas
numpy
torch
ray-2.2.0
tensorflow
torchvision